requirements.txt
# Core dependencies for TIGAS
# For development with pinned versions, use: pip install -r requirements-dev.txt

torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
huggingface-hub>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0
